pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "charter-governance"
version = "0.5.0"
description = "AI governance layer. Local-first. Open source. Three layers: hard constraints, gradient decisions, self-audit."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
daemon = [
    "flask>=3.0",
    "psutil>=5.9",
]
test = [
    "pytest>=7.0",
]

[project.scripts]
charter = "charter.cli:main"

[project.urls]
Homepage = "https://github.com/germpharm/charter"
Issues = "https://github.com/germpharm/charter/issues"
Documentation = "https://github.com/germpharm/charter#readme"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
charter = [
    "templates/*.yaml",
    "web/templates/*.html",
    "web/static/*.css",
]
